+++ src/Common/AssemblyInfo.cs
@@ -4,17 +4,17 @@ using System.Reflection;
 
 [assembly: AssemblyProduct("FSharp.Formatting")]
 [assembly: AssemblyDescription("A package of libraries for building great F# documentation, samples and blogs")]
-[assembly: AssemblyVersion("3.0.0")]
-[assembly: AssemblyFileVersion("3.0.0")]
-[assembly: AssemblyInformationalVersion("3.0.0-beta13")]
+[assembly: AssemblyVersion("3.1.0")]
+[assembly: AssemblyFileVersion("3.1.0")]
+[assembly: AssemblyInformationalVersion("3.1.0")]
 [assembly: AssemblyCopyright("Apache 2.0 License")]
 namespace System {
     internal static class AssemblyVersionInformation {
         internal const System.String AssemblyProduct = "FSharp.Formatting";
         internal const System.String AssemblyDescription = "A package of libraries for building great F# documentation, samples and blogs";
-        internal const System.String AssemblyVersion = "3.0.0";
-        internal const System.String AssemblyFileVersion = "3.0.0";
-        internal const System.String AssemblyInformationalVersion = "3.0.0-beta13";
+        internal const System.String AssemblyVersion = "3.1.0";
+        internal const System.String AssemblyFileVersion = "3.1.0";
+        internal const System.String AssemblyInformationalVersion = "3.1.0";
         internal const System.String AssemblyCopyright = "Apache 2.0 License";
     }
 }