+++ src/Common/AssemblyInfo.fs
@@ -4,16 +4,16 @@ open System.Reflection
 
 [<assembly: AssemblyProductAttribute("FSharp.Formatting")>]
 [<assembly: AssemblyDescriptionAttribute("A package of libraries for building great F# documentation, samples and blogs")>]
-[<assembly: AssemblyVersionAttribute("3.0.0")>]
-[<assembly: AssemblyFileVersionAttribute("3.0.0")>]
-[<assembly: AssemblyInformationalVersionAttribute("3.0.0-beta13")>]
+[<assembly: AssemblyVersionAttribute("3.1.0")>]
+[<assembly: AssemblyFileVersionAttribute("3.1.0")>]
+[<assembly: AssemblyInformationalVersionAttribute("3.1.0")>]
 [<assembly: AssemblyCopyrightAttribute("Apache 2.0 License")>]
 do ()
 
 module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyProduct = "FSharp.Formatting"
     let [<Literal>] AssemblyDescription = "A package of libraries for building great F# documentation, samples and blogs"
-    let [<Literal>] AssemblyVersion = "3.0.0"
-    let [<Literal>] AssemblyFileVersion = "3.0.0"
-    let [<Literal>] AssemblyInformationalVersion = "3.0.0-beta13"
+    let [<Literal>] AssemblyVersion = "3.1.0"
+    let [<Literal>] AssemblyFileVersion = "3.1.0"
+    let [<Literal>] AssemblyInformationalVersion = "3.1.0"
     let [<Literal>] AssemblyCopyright = "Apache 2.0 License"