+++ RELEASE_NOTES.md
@@ -1,3 +1,10 @@
+## 3.1.0
+  - remove beta tag since it is already widely used
+
+## 3.0.0-beta14 
+ - Update to latest FSharp.Compiler.Service
+ - No longer filter FSHarp.Core based on optdata/sigdata (it is now always bundled)
+ 
 ## 3.0.0-beta13 
  - FSharp.Literate for netstandard2.0